+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # ChangeLog
 
+- rename PreparedStatement#bind_boolean to PreparedStatement#bind_bool
 - add DuckDB::Connection#appender
 
 # 0.2.6.0